+++ modules/device/iso/default.nix
@@ -1,4 +1,4 @@
-{ ... }:
+{ inputs, ... }:
 {
   # Custom Install Media Configuration Modules
   imports =
@@ -8,5 +8,8 @@
     ../scripts
     ./packages
     ./user
+
+    # Install Media Build Module
+    "${inputs.nixpkgs}/nixos/modules/installer/cd-dvd/iso-image.nix"
   ];
 }