+++ shells/default.nix
@@ -1,0 +1,20 @@
+{ pkgs, ... }:
+pkgs.mkShell
+{
+  # Default Developer Shell
+  name = "devShell";
+
+  # Required Packages
+  buildInputs = with pkgs;
+  [
+    git
+    git-crypt
+    gnupg
+  ];
+
+  # Init Script
+  shellHook =
+  ''
+    echo "## Default Developer Shell ##"
+  '';
+}